MAlaria_app/requirements.txt
absl-py==0.13.0
astor==0.8.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dataclasses==0.8
Flask~=2.2.2
gast==0.2.2
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.38.0
h5py==2.10.0
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
numpy~=1.24.1
oauthlib==3.1.1
opt-einsum==3.3.0
pickle-mixin==1.0.2
Pillow~=9.0.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.21.3
scipy==1.5.4
six==1.16.0
tensorboard==2.1.1
tensorflow~=2.11.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1

WTForms~=3.0.1
matplotlib~=3.6.2
keras~=2.11.0